Add support for deferred URC
There is a chance that we may receive a URC during command execution.
In this case we can't call the URC callback right away since the URC callback
may execute a new AT command colliding with current command. For this reason
if a URC is received during command execution, we copy the URC line to a
separate URC buffer and after command execution is done, we trigger the URC
callback with the data in this buffer.

Please note that we will only buffer ONE URC line at the moment.